dag => index
