+++ index.ts
@@ -1,8 +1,3 @@
-export namespace flatten {
-    export type Func<I, O> = (value: I) => O[];
-    export function identity<T>(value: T): T[] { return [value]; }
-}
-
 export function lazy<T>(f: () => T): () => T {
     let called = false;
     let result: T;
@@ -13,6 +8,11 @@ export function lazy<T>(f: () => T): () => T {
         }
         return result;
     };
+}
+
+export namespace flatten {
+    export type Func<I, O> = (value: I) => O[];
+    export function identity<T>(value: T): T[] { return [value]; }
 }
 
 export namespace array {
@@ -287,5 +287,45 @@ export namespace optimized {
 
     export function one<T>(id: string, value: T): Bag<T> {
         return new Node(id, <R>(visitor: NodeVisitor<T, R>) => visitor.one(value)).bag();
+    }
+}
+
+export namespace dag {
+
+    export class Dag {
+
+        private readonly _map: { [id: string]: any } = {};
+
+        get<T>(bag: bag.Bag<T>): optimized.Bag<T> {
+            const id = bag.id;
+            const cached = this._map[id];
+            if (cached !== undefined) {
+                return cached;
+            }
+            const getOpimized = <I>(b: bag.Bag<I>) => this.get(b);
+            class Visitor implements bag.Visitor<T, optimized.Bag<T>> {
+                flatten<I>(value: bag.Flatten<T, I>): optimized.Bag<T> {
+                    return getOpimized(value.input).flatten(id, value.func);
+                }
+                disjointUnion(value: bag.DisjointUnion<T>): optimized.Bag<T> {
+                    return getOpimized(value.a).disjointUnion(id, getOpimized(value.b));
+                }
+                one(value: T): optimized.Bag<T> {
+                    return optimized.one(id, value);
+                }
+                input(): optimized.Bag<T> {
+                    return optimized.input<T>(id);
+                }
+                groupBy(value: bag.GroupBy<T>): optimized.Bag<T> {
+                    return getOpimized(value.input).groupBy(id, value.toKey, value.reduce);
+                }
+                product<A, B>(value: bag.Product<T, A, B>): optimized.Bag<T> {
+                    return getOpimized(value.a).product(id, getOpimized(value.b), value.func);
+                }
+            }
+            const result = bag.implementation(new Visitor());
+            this._map[id] = result;
+            return result;
+        }
     }
 }
